Refactored relative_eq
